workflows/clean/pyproject.toml
[project]
name = "clean"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dagger-io",
    "email-validator>=2.2.0",
    "pydantic-ai>=0.2.4",
    "pyyaml>=6.0.2",
    "simple-chalk>=0.1.0",
    "supabase>=2.15.1",
]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
